Add script for checking header files with clang (#113)
* Always use cpp extension for C++ files

* Use standard C for MMIO for non-mwcc

* Add missing include in sb_ascii.h

* Remove useless types.h include in some dwc headers

* Fix capitalization of some dwc includes

* Only include eabi header if mwcc

* Deduplicate GXRenderModeObj definition

* Fix some C stdlib declarations

* Add typedef for structures used without namespace

* Disable asm functions for non-mwcc

* Always use unsigned long for size_t

* Deduplicate some IOS functions

* Disable section declspec for non-mwcc

* Skip C stdlib functions with va_list for non-mwcc

They cause issues with clang.

* Fix position of override modifier in eggColorFader

* Add a missing override modifier

* Fixup some not yet included headers

Include paths mainly.

* Use MWREG macro instead of register

* Disable more asm functions

* Disable restrict in C++ mode

* Handle never_inline on non-mwcc

* Add script for checking header files with clang

* Add Windows support to clang_check.py

* Use MW_PRAG macros in math.h

* Disable MARK_BINARY_BLOB on non-mwcc

* Remove unused pragma

* Add -Wall to clang_check.py

* Remove const qualifier in return type

* Fix signed to unsigned comparison

* Add -Wextra to clang_check.py

* Always put a new line at end of file

* Remove useless semicolons

* Do not overflow int in enum values

* Stop using anonymous structs in C++

* Disable flexible array members in C++ structs

* Remove unused define

* Add -Wpedantic to clang_check.py

* Add -Werror to clang_check.py

* Add clang_check.py to CI

+++ source/platform/wchar.h
@@ -4,7 +4,7 @@
 
 typedef unsigned short wchar_t;
 
-u32 wcslen(const wchar_t*);
+size_t wcslen(const wchar_t*);
 
 wchar_t* wcscpy(wchar_t*, const wchar_t*);
